Revert "Plan 9 from Bell Labs Fourth Edition 2007-01-22"
This reverts commit e80b5be60f91cbf481a25b864b3470413b01880a.

+++ sys/src/9/port/devenv.c
@@ -262,7 +262,7 @@ static long
 envwrite(Chan *c, void *a, long n, vlong off)
 {
 	char *s;
-	ulong vend;
+	int vend;
 	Egrp *eg;
 	Evalue *e;
 	ulong offset = off;
